Refactored dmarc report summary
