+++ api/tests/testdata/dmarc_report_summary/__init__.py
@@ -1,0 +1,3 @@
+from tests.testdata.dmarc_report_summary.dmarc_report_summary_data import (
+    api_return_data,
+)
